Up to Ep 18 - pick up items and have UI pop ups

+++ Assets/Scripts/PlayerStats.cs
@@ -12,7 +12,7 @@ namespace CB_DarkSouls
 
         public int staminaLevel = 10;
         public int maxStamina;
-        public int currentStamina;
+        public float currentStamina;
 
         public HealthBar healthBar;
         public StaminaBar staminaBar;
